move demo for consistency
